added PUT / DELETE to comments, Query stock with page limit

+++ api/Controller/StockController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using api.Data;
 using api.Dtos.Stock;
+using api.Helpers;
 using api.Interfaces;
 using api.Mappers;
 using api.Models;
@@ -24,15 +25,19 @@ namespace api.Controller
             dbContext = DBContext;
         }
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query)
         {
-            List<Stock> stocks = await stockRepository.GetAllAsync();
+            if (!ModelState.IsValid)
+                return BadRequest();
+            List<Stock> stocks = await stockRepository.GetAllAsync(query);
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stockDto);
         }
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
             Stock? stock = await stockRepository.GetByIdAsync(id);
             if (stock == null)
             {
@@ -43,14 +48,18 @@ namespace api.Controller
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
             Stock? stockModel = stockDto.ToStockFromCreateDto();
             await stockRepository.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
             Stock? stockModel = await stockRepository.UpdateAsync(id, updateDto);
             if (stockModel == null)
             {
@@ -59,11 +68,15 @@ namespace api.Controller
             return Ok(stockModel.ToStockDto());
         }
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
             Stock? stockModel = await stockRepository.DeleteAsync(id);
-            return NoContent();
+            if (stockModel == null)
+            { return NotFound(); }
+            return Ok(stockModel.ToStockDto());
         }
     }
 }
